Support basic re-ordering in rebase dashboard.
It was necessary to add a `git reset --hard` at the end of RewriteMixin#rewrite_active_branch.
If a conflict occurred while cherry-picking, notify the user.  They could then
drop to the terminal if necessary to make the changes using `git rebase -i`.
If this is an issue in enough cases where re-ordering would normally be used,
the issue should be revisited.  Visual indicators could be incorporated into
the rebase dashboard similarly to what is planned for conflicts during actual
rebases.